Dubins factory functions
